Add new port comms/dabstick-radio:
While preparing for a release 4.2 of the sdr-j software package,
we created two new programs

    a DAB receiver
    a broad spectrum version of the FM receiver

It is quite obvious that one can use a DAB stick to receive DAB
programs. However, the DAB software presented here is a real SDR
in that it uses the 8 bit I/Q samples of the DAB stick.

WWW: http://www.sdr-j.tk/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@316735 35697150-7ecd-e111-bb59-0022644237b5