Apply several upstream svn revisions that have also been merged to the base version:
r170353:
Fix another SROA crasher, PR14601.

This was a silly oversight, we weren't pruning allocas which were used
by variable-length memory intrinsics from the set that could be widened
and promoted as integers. Fix that.

r175057:
X86: Disable generation of rep;movsl when %esi is used as a base pointer.

This happens when there is both stack realignment and a dynamic alloca in the
function. If we overwrite %esi (rep;movsl uses fixed registers) we'll lose the
base pointer and the next register spill will write into oblivion.

Fixes PR15249 and unbreaks firefox on i386/freebsd. Mozilla uses dynamic allocas
and freebsd a 4 byte stack alignment.

r175360:
MCParser: Reject .balign with non-pow2 alignments.

GNU as rejects them and there are configure scripts in the wild that check if
the assembler rejects ".align 3" to determine whether the alignment is in bytes
or powers of two.

r175962:
X86: Disable cmov-memory patterns on subtargets without cmov.

PR:		ports/176269, ports/176893, ports/176967
Requested by:	tijl, dim, others


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@318798 35697150-7ecd-e111-bb59-0022644237b5