Welcome SDL2 to the FreeBSD ports system!
A set of new SDL2 related ports has been added, which are essentially
the same as their corresponding port for SDL1.2:

    audio/sdl2_mixer
    graphics/sdl2_image
    graphics/sdl2_ttf
    net/sdl2_net

A set of new USE_SDL knobs is available to enable SDL2 and its related
ports to be added as dependency to ports. Use either of

  USE_SDL= sdl2 image2 mixer2 net2 ttf2

to pull in the specific sdl2_* or sdl20 port.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@321500 35697150-7ecd-e111-bb59-0022644237b5