Update the KDE Software Compilation to 4.10.5.
Proudly brought to you by the KDE on FreeBSD team. We're sorry to ship two
KDE updates in just a few days, but the work on 4.10.5 was very light
compared to 4.10.4 so it was ready much faster.

The release announcement can be found in [1].
[1] http://www.kde.org/announcements/announce-4.10.5.php

The biggest news for us on FreeBSD is that the Ruby bindings should work
with Ruby 1.9 now.

I will probably add a note to UPDATING later about this, but as avilla@
pointed out, the clang support we mentioned that was improved in 4.10.4
requires a rebuild of the ports that depend on kdelibs4. Most of them are
covered by this update, but those which are not part of the Software
Compilation need to be rebuilt manually to make sure the previous issues
(proper symbol visibility being the most annoying of them) are solved.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@322300 35697150-7ecd-e111-bb59-0022644237b5