devel/py-paver: easy build, distribution and deployment scripting
Paver is a Python-based software project scripting tool along the lines of Make
or Rake. It is not designed to handle the dependency tracking requirements of,
for example, a C program. It is designed to help out with all of your other
repetitive tasks (run documentation generators, moving files about,
downloading things), all with the convenience of Python's syntax and massive
library of code.

WWW: http://pythonhosted.org/Paver/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@322941 35697150-7ecd-e111-bb59-0022644237b5