Add new port: lang/asis
This is ASIS, the Ada Semantic Interface Specification.
ASIS is tightly coupled with lang/gcc-aux and must be built with the same
sources.  For this reason, ASIS will not work with other Ada compilers,
including GNAT GPL (lang/gnat).

ASIS is a dependency of several upcoming Ada ports.

Approved by:	bapt (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323389 35697150-7ecd-e111-bb59-0022644237b5