security/libsparkcrypto: Optionally generate documentation
It was not possible to generate libsparkcrypto documentation before the
textproc/adabrowse port was created.  Now that AdaBrowse is available,
add a "DOCS" options to generate the documentation.

While here, convert USE_GMAKE to USES+=gmake and fix DESTDIR to work
properly instead of installing during the build phase.  The Makefile
also had to altered to install the documentation as well.

Approved by:	bapt (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323394 35697150-7ecd-e111-bb59-0022644237b5