games/freetennis: Adjust dependencies
There were a few problems with the dependency definitions of freetennis.
The ocaml-images dependency is a static build dependency, and not a RUN
DEPENDS.  Secondly, a specific static library is specified, but no
library of that name exists, at least not for the upcoming ocaml-images
4.0.  Freetennis really wants the entire directory, so ocaml-images has
been changed to a package depends instead of a file depends.  It's been
tested with ocaml-images 4.0 (DragonFly) and ocaml-images 3.0 (FreeBSD/
Redports).

Approved by:	bapt/culot (mentors, implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@324890 35697150-7ecd-e111-bb59-0022644237b5