science/minc2: Switch license to BSD
The license was defined as "AS-IS" but this license has no definition in
bsd.license*.mk.  As a result, this port would not build in bulk because
there was no implied acceptance of the "AS-IS" license.  Upon
re-examination, the maintainer believes it is close enough in spirit to
BSD license to be classified as BSD.

   Copyright 1993-2000 Peter Neelin and David MacDonald, McConnell Brain
   Imaging Centre, Montreal Neurological Institute, McGill University.
   Permission to use, copy, modify, and distribute this software and its
   documentation for any purpose and without fee is hereby granted,
   provided that the above copyright notice appear in all copies. The
   author and McGill University make no representations about the
   suitability of this software for any purpose. It is provided "as is"
   without express or implied warranty.

The PORTREVISION was bumped because when a license is specified, it
modifies the pkg-plist.

PR:		181350
Approved by:	bapt/culot (mentors, implicit), jwbacon (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@325245 35697150-7ecd-e111-bb59-0022644237b5