Work around a bug in libcpp that pulls in the optional system iconv if present instead of relying on the port we actually depend on.
The issue is that /usr/include/iconv.h has #include <stdbool.h> which
in turn, since both are included very late in the game, conflicts with
similar definintions by libcpp itself.

Interestingly enough, lang/gcc46 and lang/gcc48 did not fail under the
same test scenarios and libstdc++ does not seem to require adjustments,
so PR 161417 may not be relevant in full.

PR:		161417


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@326441 35697150-7ecd-e111-bb59-0022644237b5