mail/textmail: Fix previous pdftotext breakage better
This unmaintained port was marked broken after pdftotext was relocated
away from bin/ with the rest of the xpdf package.  That happened to
avoid a conflict with poppler-utils.  The better fix is to change the
dependency to use pdftotext program from the preferred poppler-utils
port which remains installed in the standard bin/ path.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@329638 35697150-7ecd-e111-bb59-0022644237b5