Explicitly link libraries on 10 unmaintained ports
The following ports will not build with a binutils 2.22+ linker built
with standard options.  This has been obvious with DPorts, but difficult
to see on FreeBSD.  However, setting the ports compiler as a recent gcc
(e.g lang/gcc48) is an excellent way to detect the unspecified but needed
libraries as these recent GCC compilers use the latest binutils linkers.
These patches were tested on FreeBSD 8.4 and DragonFly 3.5


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@329743 35697150-7ecd-e111-bb59-0022644237b5