databases/gtksql: Fix leftovers
Ever since this port was updated to verson 0.4.5, it has suffered from
leftover files caused by installing two sets of PORTDOCS.  The custom
version added in the post-install target is accounted for, but the
distfiles's makefile also has a target that installs the same files
plus two others in a non-standard location.  The fix is to disable
the distfile's install target and keep the one in post-install.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@329773 35697150-7ecd-e111-bb59-0022644237b5