STAGIFY
Allow to build on amd64 and ia64 since pvm has had shared libs on all
platforms for quite some time.

Mark DEPRECATED as it seems unlikely there are any actual users given
that it hasn't worked on amd64 ever and the upstream hasn't made any
changes for 8 years or so.

Drop maintainership.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@330753 35697150-7ecd-e111-bb59-0022644237b5