Disable the default dependency on LAME for xbmc so that package building and distribution becomes possible.
LAME is a restricted package so we cannot ship a package that depends on it.
According to the maintainer it is only used for ripping CDs, and there are
better tools than XBMC to do just that.

Also fix the build with a C11 compiler (clang). There's now a weak_ptr in
boost::, and one in std::.

Approved by:	maintainer


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@331589 35697150-7ecd-e111-bb59-0022644237b5