archivers/lzip: Unbreak fetch with LOCAL
The lzip developers do something very annoying.  When they issue a new
release, they remove the previous release's tarball in .gz format.
They keep the tarball in .lz format, but move it to the "old"
subdirectory.  The result is that gzipped tarballs for lzip get very
scarce, and requiring lz to unzip lz is not an option.  I don't
normally store distfiles for other maintainers, but it appears that
many folks have been asking for this port to be fixed, so I made
an exception and put a copy of lzip tarball in LOCAL/marino.

PR:		ports/182982
Approved by:	maintainer timeout


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@332321 35697150-7ecd-e111-bb59-0022644237b5