Mk/bsd.database.mk: Fix postgresql component specifications
Due to the RUN_DEPENDS and BUILD_DEPENDS specification of postgresql
components in bsd.database.mk not matching their package names,
net-mgmt/netmagis-database was failing during the RUN_DEPENDS stage
because it couldn't detect that a component was already installed.

Updating the pltcl line allowed that port to successfully built.  The
other lines weren't tested but the problem would be the same.

Suggested by:	wg


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@332609 35697150-7ecd-e111-bb59-0022644237b5