Make port option less confusing.
In the context of this port, turning on yaml support
doesn't enable the optional yaml support like it does
for other ports.  It compiles the yaml support against
textproc/libyaml, which provides for yaml parsing with
the speed of C in the ease of use of a scripting language.

PR:	ports/183763
Submitted by:	Michael Gmelin <freebsd@grem.de>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@333337 35697150-7ecd-e111-bb59-0022644237b5