New port: mail/p5-Email-Stuffer:
Email::Stuffer, as its name suggests, is a fairly casual module used to stuff
things into an email and send them. It is a high-level module designed for
ease of use when doing a very specific common task, but implemented on top of
the light and tolerable Email:: modules.

Email::Stuffer is typically used to build emails and send them in a single
statement, as seen in the synopsis. And it is certain only for use when
creating and sending emails. As such, it contains no email parsing
capability, and little to no modification support.

WWW: http://search.cpan.org/dist/Email-Stuffer/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@334546 35697150-7ecd-e111-bb59-0022644237b5