ports-mgmt/pkg*: Prepare for alternative PORTSDIR support
With the PKG_PORTSDIR make argument set, Release 1.2.2 will properly build
pkg when PORTSDIR doesn't equal the default /usr/ports, as is the standard
case with DragonFly.

Additionally, pkg(8) and pkg-static(8) man pages are conditionally edited
to remove references to the -j option when libjail isn't found on the
system.  This extra-patch should never be used on FreeBSD.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335210 35697150-7ecd-e111-bb59-0022644237b5