math/py-numpy: Fix static library check that assumes atlas is installed
- Use BUILD_DEPENDS to check for the atlas static library at pre-build
  time instead of checking if the file exists (False if atlas is not
  installed)

- Replace IGNORE with a message to the user notifying them that
  an atlas install WITH_STATIC is needed, and provide clues as possible
  actions without being too prescriptive.

As discussed with eadler@, a correct resolution to "ports depending on other ports with specific OPTIONS" is either an atlas slave port
(ie: atlas-static), or to modify atlas to install both shared and static
libraries.

This change is a net gain on the existing broken behaviour, by fixing
the check, while retaining the user notification.

PR:		ports/178094
Reviewed by:	jhibbits, eadler
Approved by:	maintainer timeout (7+ months)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@336423 35697150-7ecd-e111-bb59-0022644237b5