Add x11-fonts/vollkorn-ttf.
It intends to be a quiet, modest and well working text face for bread and
butter use. Unlike its examples in the book faces from the renaissance until
today, it has dark and meaty serifs and a bouncing and healthy look. It might
be used as body type as well as for headlines or titles.

WWW: http://friedrichalthausen.de/vollkorn/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@337372 35697150-7ecd-e111-bb59-0022644237b5