games/daimonin-music: Remove NO_CHECKSUM
It appears the original maintainer was trying to have the port
continue to work even when the AllMusic.zip distinfo file got updated.
He also worked under the assumption that the port was updated in place
rather than be deinstalled and reinstalled, hence all the logic in the
pkg-plist previously.

I removed NO_CHECKSUM, generated a distinfo file, and updated the
portversion to match the latest modification date (2012-04-02).  The
related PR provided minimal inspiration, but the same goal has been
achieved.

PR:		ports/170052


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@337630 35697150-7ecd-e111-bb59-0022644237b5