shells/bash-completion: Fix installation
The bash-completion.pc and bash_completion.sh files were not getting
installed in the correct location per pkg-plist, mainly because the
Makefile.in was modified *after* the configuration target was complete,
that is to say after Makefile has been generated.

By improving the Makefile.am patch, the post-configure target can be
simplified.

Approved by:	maintainer (adamw)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@337881 35697150-7ecd-e111-bb59-0022644237b5