Fix installation for slave port of emulators/wine-devel.
Wine switched to using an internal install script and removed logic to create
the underlying directories.  As explained in a comment from Makefile:

"""
Wine expects the install script to make any missing directories, and the default
internal tool (/tools/install-sh) does that however the default install script used
by Ports does not.
"""

Disable the specification of the INSTALL scripts and allow wine to use the
internal one.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@339006 35697150-7ecd-e111-bb59-0022644237b5