converters/fconv: Unbreak for root builds
The previous commit enabled non-root builds to work, but broke builds
by root instead.  The ${GID} variable is blank when the user is root,
so the replacement of "root" with ${GID} resulted in bad command.
Just use ${BSD_INSTALL_PROGRAM} instead and revbump because the packaged
binaries could be stripped now because of that.  Commit covered by the
"Just fix it" blanket.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@339621 35697150-7ecd-e111-bb59-0022644237b5