Update to the 20140209 snapshot of GCC 4.9.0.
Configure with --enable-gnu-indirect-function to enable the "ifunc"
attribute, a GCC extension. [1]

Remove the man7 pages, which we are not packaging, from the staging area
to silence some warnings some are concerned about (needlessly ;-). [2]

No longer run ccache-update-links as part of post-install which, in
the world of staging, no longer is what it used to be.  Rely on the
existing @exec and @unexec in pkg-plist instead. [3]

Submitted by:	jmg [1], miwi [2]
Discussed with:	kib [1], antoine [2][3]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344030 35697150-7ecd-e111-bb59-0022644237b5