devel/adabooch: Rework installation target
For some reason, $prefix was already defined on DragonFly but
I didn't realize this was the reason it wouldn't install there until
after completely revamping the installation target.  In any case,
it's much better now.  No bump as there are no difference for FreeBSD
and it never built on DragonFly.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@345092 35697150-7ecd-e111-bb59-0022644237b5