No longer add -I${LOCALBASE}/include to CFLAGS.
Since we now configure with --with-gmp=${LOCALBASE} this is no longer
necessary, and due to bugs in binutils (which should not install ansidecl.h
into ${PREFIX}/include, fixed with revision 336642 [1]) and GCC (which
should search its own include directories with higher priority) could
lead to build failures.

Set the license to a combination of GPLv3 (for the compiler itself) and
GPLv3 with GCC Runtime Library Exception (for the runtime).

PR:		184327 [1]


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@345671 35697150-7ecd-e111-bb59-0022644237b5