devel/fib: Honor NO_PROFILE, fix staging and override install target
This port was building and installing the profile library regardless of
the NO_PROFILE definition.  The installation was a bit of a mess too, and
it didn't use the BSD_INSTALL_* values.  The previous staging attempt
didn't work, at least not on DragonFly, so let's use our own install
target to fix all of these issues and honor the NO_PROFILE request
at the same time.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346717 35697150-7ecd-e111-bb59-0022644237b5