devel/py-unittest2: Update LICENSE w/ references
- Update LICENSE (BSD3CLAUSE)

This one was fun to find. As rm@ mentioned in a previous commit, no
LICENSE file exists in the Google Code or hg.python.org repositories.

There is a "New BSD License" reference on the unittest-ext homepage [1]
but wasnt a direct reference to the source in question, so it was not
considered good enough.

Gold was struck grepping BSD in the sources, which leads to the setup.py
header, that includes a URL reference to:

    http://www.voidspace.org.uk/python/license.shtml

BSD 3-clause. Voila!

[1] http://code.google.com/p/unittest-ext/

Approved by:	portmgr (blanket)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@347449 35697150-7ecd-e111-bb59-0022644237b5