Duplicate port lang/gcc-aux => lang/gcc47-aux
A new version of the Ada compiler is coming imminently.  Not all ports
will immediately build with the gcc49-based compiler, so a few ports
will be directed to use the current compiler in this new location instead.
It will require a change in Uses/ada.mk (upcoming).

Some ports will require updates to use upcoming lang/gcc-aux version.
The last set of ports that will be converted is likely the gnatdroid
ports as these will require a lot of time running testsuites, which takes
some time wirelessly on the Android target.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@347679 35697150-7ecd-e111-bb59-0022644237b5