net/adasockets: Fix code style before gcc49 comes in
Part of adasockets is built with -gnatg which invokes style checking.
The upcoming GNAT compiler has stricter rules regarding the placement
of the "then" keyword.  Adasockets was failing to build based on style
errors on gcc49.  Patch this now on gcc47 to prepare for the Ada
compiler transition.

No bump; the code logic doesn't change.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@347972 35697150-7ecd-e111-bb59-0022644237b5