- Update to gnuradio 3.7.2.1
GRC Bus Ports (Nicholas McCarthy)

Bus ports in the GNU Radio Companion (GRC) allow ganging block input
or output ports into a single port for display and connections. This
simplifies flowgraph entry when dealing with blocks that work with an
arbitrary number of streams.

For further information, see:

http://gnuradio.org/redmine/projects/gnuradio/wiki/Busports

And see

http://gnuradio.org/redmine/projects/gnuradio/wiki/ChangeLogV3_7_1


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@349385 35697150-7ecd-e111-bb59-0022644237b5