Release maintainership.  There is new version available, 0.5.4, but it does not work, exiting with "Trace/BPT trap" error on startup and dumping core.
Our devd(8) support patch was submitted upstream in July 2013 and generally
welcomed by the author, but he has a slight preference for GIO's GSocket +
GUnixSocketAddress APIs, and I never had time to rewrite the patch.  After
showering my camera with waters of Gulf of Thailand, it is unlikely that I
will need to use this port any time soon, so let it go for the moment.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@350585 35697150-7ecd-e111-bb59-0022644237b5