games/minecraft-client: Unbreak run by using official launcher
The method of selectively loading specific versions in Minecraft stopped
working in the 1.5.x series.  From 1.6.0 and later, the official launcher
must be used.  The launcher itself does not change often (last time was
July 2013).  The jar files are downloaded into $HOME/.minecraft directory,
so every user gets their own copy which is not removed when the port is
uninstalled.

The new launcher also downloads lwjgl files, so the dependency on
games/lwjgl is redundant, as well as openal-soft.  This allows the port
to be simplified and to always install the native version.

The port version doesn't really matter anymore.  I will set it to 1.7.9
right now as this is the latest available version.

PR:		ports/188473
Approved by:	marino@. It does mirror the maintainers original intent,
		but he did not weigh in on the PR.  The game was broken.
Simplified by:	marino@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351506 35697150-7ecd-e111-bb59-0022644237b5