Welcome mod_cloudflare to the tree.
Unfortunately, Cloudflare doesn't seem to be handling release management
of mod_cloudflare very well so for now we are pulling from their public
Github repository and the version number will be 0.0.YYYYMMDD based on
the date of the commit we're pulling. They don't commit updates very
often, so this should be fine until they figure out their release
process.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351951 35697150-7ecd-e111-bb59-0022644237b5