x11-toolkits/p5-Alien-wxWidgets: Unbreak on some platforms
For me, the most cringeworthy makefile technique is the shell command
(!=) evaluating another port with "make -C".  This is very slow, fragile,
and can cause a huge mess if something is out of wack.

The technique was used in this port and it didn't even work on all
platforms.  For sure FreeBSD 8 (according to portsmon) and DragonFly
were broken.  Essentially the file "gtk2_3_0_0_uni_gcc_3.4.pm" was
installed, but the broken plist substitution was looking for
"gtk2__uni_gcc_3.4.pm".

To fix this robustly, get rid of the whole "make -C" mess and the plist
substituion and just use the find command to dynamically add this one
file to the plist.  Since it starts with a unique pattern, we really
don't need to know the Wx version in order to establish the plist with
this technique.

Approved by:	blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@352448 35697150-7ecd-e111-bb59-0022644237b5