- Add devel/libstatgrab0 (copied from devel/libstatgrab) - Switch bindings to use devel/libstatgrab0 instead of devel/libstatgrab - Fix devel/pecl-statgrab to work with current PHP versions
I have an update ready for devel/libstatgrab (version 0.90) that
brings a number of API changes that are incompatible with the current
version. These bindings haven't yet been updated upstream to support
the new API, so instead of holding off indefinitely I've created the
devel/libstatgrab0 port to provide just the library and header file
for the older API version so that these bindings can continue to be
used when the main port is updated. Once these bindings are updated
the libstatgrab0 port can be removed.

Both devel/libstatgrab and devel/libstatgrab0 can be installed at the
same time.

The other ports that depend on devel/libstatgrab will be updated to
use the new API.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@353876 35697150-7ecd-e111-bb59-0022644237b5