net/anet: Upgrade version 0.2.3 => 0.3.0
This version fixes all the stuff the removed patches addressed (this
version was released as a direct result of these patches which I provided
upstream).  Multicast sending remains tricky, in part due to a known and
yet to be addressed bug in FreeBSD.

The pkg-message was deleted.  I don't completely understand what needs
to be established in order for IPv6 multicast to work (e.g. don't use
the same machine for sending and receiving) so rather than mislead, I
just removed this information for now.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@354028 35697150-7ecd-e111-bb59-0022644237b5