devel/gprbuild: Upgrade version 20130416 (2013) => 20140417 (2014)
Adacore provides annual releases for its major GPL-licensed products.
GPRBuild is an important tool, but it's source is only exposed once
per year, in May.  The 2014 release of GPRBuild relies on features in
Adacore's commercial product, GNAT Pro, that has not yet been migrated
to FSF GNAT.  I had to patch 2014 to get it to build with FSF GNAT 4.9,
losing minor new functionality in the process.

My first decision was to simply skip GPRBuild 2014 stay on the 2013
version, but it turns out packages are already using the new features
in gprinstall, one of the tools in the GPRBuild package.

Tests on adabooch, aws, asis and opentoken indicate they build fine
with GPRBuild 2014.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@354203 35697150-7ecd-e111-bb59-0022644237b5