devel/gnatcoll4: Rename gnatinspect and toggle off iconv option default
The GNAT Programming Studio wants to use gnatinspect, so it needs to
build it.  At the same time, gnatinspect also belongs to gnatcoll.
Resolve a filename conflict by renaming it to gnatinspect-xref when
it's build by gnatcoll.

Also turn off iconv support by default.  It causes GPS to crash
almost immediately and the issue is likely within the gnatcoll code.
Until this is investigated and hopefully fixed, keep it off by default.

Also, devel/gnatcoll will not be a dependency of GPS for two reasons:
GPS is so fluid that it really needs the embedded version to guarantee
that it can be built, and secondly devel/gnatcoll and devel/gps could
easily need different build options.  So that leaves devel/gnatcoll as
purely a standalone development library.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355092 35697150-7ecd-e111-bb59-0022644237b5