lang/gnat_util: Carry impunit.ad[bs] source files
The GNAT Programming Studio needs the impunit.ad[bs] files in order
to create a menu based on the gnat source code.  Just carry these
source files with the others that are used to build gnat_util and then
use the GPS port makefile to pull it in as needed.  A quick check with
poudriere indicates that ports dependent on gnat_util still build fine.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355093 35697150-7ecd-e111-bb59-0022644237b5