Makefile.common: Always set USES=tar:bzip2.
Setting USES=bzip2 inside the NO_BUILD block caused problems for FreeBSD's
distfile fetcher, which apparently set that flag and was getting an invalid
distfile to fetch from the KDE mirrors.

The rest of the NO_BUILD block needs more attention, but I'll let makc@ and
jhale@ handle that since it's not my area.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355452 35697150-7ecd-e111-bb59-0022644237b5