devel/cvs2svn: Replace EXTRACT_CMD with TAR
This port has distfiles of two formats: .tar.bz2 and .tar.gz.
The TAR command handles both of them, so explicit specify this tool
to ensure the port continues to work in the future.

As a comment, it's strange that a custom extract target is needed to
overcome a possible deficiency in the standard extract target.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355987 35697150-7ecd-e111-bb59-0022644237b5