www/templatelite: Stage support and cleanup
The suggested patch was technically correct, but it was putting lipstick
on a pig.  The install target was pretty ugly.  Let's clean this up a bit
without homegrown copytrees, then pet portlint.

Also, get rid of NOPORTDOCS in favor of standard options.  Sorry, not of
the original patch is left.

PR:		ports/190683
Submitted by:	joe@thrallingpenguin.com
Reworked by:	marino


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@356787 35697150-7ecd-e111-bb59-0022644237b5