math/why3-gpl: Increase distinction between this and math/wny3
The why3 project is worried that users will be confused between this
package and a "vanilla" why3, which was simultaneously added with this
one.  They prefer that this port be completely renamed.

While I ponder that, I can at least improve the situation by fixing the
descriptions to lessen the chance of confusion between the ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@357001 35697150-7ecd-e111-bb59-0022644237b5