Switch mail/qmail and mail/qmail-mysql to using netqmail. Switch their PKGNAMEs to reflect this as well and bump their version up to the last netqmail release of 1.06.
qmail-ldap and qmail-spamcontrol explicitly want to be applied on stock
qmail, so leave those alone.

This does not bring much, but it allows easier updates to qmail-tls and
some of the other patches which depend on netqmail.

qmail-tls is left as qmail for now. It will be updated to the newer
revision soon which does require netqmail.

Changes include:
  * QMAILQUEUE by default
  * Fix for 0.0.0.0
  * qmail-smtpd is protected from exceedingly long (eg 2GB) headers

Other than those listed, there are no other significant changes beyond
documentation or patches that were already applied by default.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@358497 35697150-7ecd-e111-bb59-0022644237b5