MFH: r360338
Fix patch application in mail/qmail-spamcontrol:

The spamcontrol patch is applied through a script, install_spamcontrol.sh. This
script does not return a non-zero status when patch failures are encountered.
So I missed many failures when updating this port.

To fix this, move the spamcontrol patch application to pre-patch so that the
PATCHFILES->DISTFILES (and manual patch application with PATCH) is not needed
for most patches. This ensures that the spamcontrol patch is always applied
before PATCHFILES. Some PATCHFILES were applying before the spamcontrol patch
causing it to fail to patch properly.

Bump revision as the build had been fine but some files were not being
generated in the package.

Also add my mirror as distcache does not have the updated patches yet.

Reported by:	Oleg Strizhak