astro/wcslib: USES+= fortran (for wcunitse)
Back in August 2013 I reported to the maintainer that if fortran is not
found then certain functions such as wcunitse are not built.  These
functions are needed for kstars which is a dependency for kdeedu and the
kde4 meta port.  The USE_FORTRAN=yes was never added though, so lets do
that now that the port is unmaintained.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@360750 35697150-7ecd-e111-bb59-0022644237b5