multimedia/kplayer-kde4: Stage, fix on F10+, use explicit linking
Apparently the clang parser was having trouble with the nested ternary
operator on C++ which was the cause of the FreeBSD 10+ failures. Expanding
it to a condition ladder pleased clang.

A couple of libraries were used but not explicitly specified to the
linker before: libsolid and libX11, so let's correct that too.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@361178 35697150-7ecd-e111-bb59-0022644237b5