New port: security/signify
'signify' from OpenBSD as of July 15 2014, with the necessary plumbing
cobbled together from the OpenBSD CVS to make it compile under Linux.

It can be used to verify the LibreSSL distfile from the OpenBSD archive.

Since the CVS itself is not signed, the author provides no guarantees.

WWW: http://www.fefe.de/signify/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@361987 35697150-7ecd-e111-bb59-0022644237b5