deskutils/calibre: Update version 1.44.0 => 1.45.0 plus clean-up
The PR as provided did not pass stage checks.  Not only were there new
installed files not represented in the pkg-plist, but patching the html
files caused the .orig versions to be installed too.  That required
a post-patch target to remove as well as the updated pkg-plist.

A long standing issue that was non-fatal on FreeBSD but fatal on DragonFly
involved the use of the calibre RC script.  The calibre.in template is
filled in at ${WRKDIR}/calibre, the same exact location for the default
configure directory.  FreeBSD handled that problem like this:

  "No write acces [sic] to /work/a/ports/deskutils/calibre/work/calibre
   using a temporary dir instead"

But the error code returned by DragonFly caused this message there:
  "OSError: [Errno 20] Not a directory: '/wrkdirs/deskutils/calibre/
   work/calibre/global.py'"

The solution was to set CALIBRE_CONFIG_DIRECTORY to another location.
Strangely this directory is removed by the build later.

Another change I made was to move XDG_CONFIG_HOME from
${STAGEDIR}/.config to ${WRKDIR}/xdg-config where it doesn't need to
be deleted as a post-install step.  Strangely, this directory is empty
after the build, so all this XDG_CONFIG_HOME can probably be removed
completely.

The remaining changes were wrapping lines to respect an 80-column limit
and unmasking install commands.

PR:		191957
Submitted by:	maintainer (Rusty Nejdl)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@362498 35697150-7ecd-e111-bb59-0022644237b5