- Add rocksdb 3.2
RocksDB is an embeddable persistent key-value store for fast storage. RocksDB
can also be the foundation for a client-server database but our current focus is
on embedded workloads.

RocksDB builds on LevelDB to be scalable to run on servers with many CPU cores,
to efficiently use fast storage, to support IO-bound, in-memory and write-once
workloads, and to be flexible to allow for innovation.

GitHub repository at https://github.com/facebook/rocksdb

WWW: http://rocksdb.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@362558 35697150-7ecd-e111-bb59-0022644237b5