x11/xcb: Fix incorrectly built app-defaults file
The app-defaults file for xcb was not built correctly as a consequence of
the switch to clang and replacement of gcc's cpp to tradcpp which is more
strict about the order of the parameters.  The filename must be last on
the command line otherwise a zero-length file is created.

Fixed by moving the filename parameter to the end of the cpp command

PR:		188203
Submitted by:	callum (omma.gibson.athome)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@363029 35697150-7ecd-e111-bb59-0022644237b5