Update to MPICH 3.1.2
From the announcement:
This release contains significant enhancements to the BG/Q device,
especially for RMA and shared memory functionality.  It also contains
enhancements to ROMIO and upgrades hwloc to 1.9.  In addition, it updates
its weak alias support to align with gcc-4.x, has a better implementation
of MPI_Allreduce for intercommunicator, adds more F08 test cases and fixes
several bugs present in 3.1.1.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@365036 35697150-7ecd-e111-bb59-0022644237b5