www/squid33: Support Kerberos on DragonFly
DragonFly does not have Kerberos in the base system as this port assumes.
In order to build squid33, DF has a dependency on krb5 packages.  Morever,
the configure check assumes Kerberos is in the base system, so it fails
to configure.  Disable the check completely as it's guaranteed to pass on
FreeBSD and the sought headers are installed with krb5.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@365726 35697150-7ecd-e111-bb59-0022644237b5