cad/alliance: Upgrade version 5.0.20090901 => 5.0.20120515, stage
This port was marked broken due to a bison error.  The newest version had
the same flaw, but I got the fix from Debian.  The 5.0.20120512 is also
not limited to i386, it builds fine on amd64 now.

It is still not jobs safe (it can't find libVrd before it's built) but
I suspect only minor changes are necessary to make it jobs safe.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366644 35697150-7ecd-e111-bb59-0022644237b5