Make 7 ports jobs unsafe
A number of ports failed on the DragonFly package builder when
theoretically all of them should build.  They had errors in logs
indicating that was parallel building issue.  Rerunning through again
with -j1 resulted in a good build.  There were more than 7, but these
seven are the ports that I'm confident are not 100% jobs safe based on
the errors seen.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366784 35697150-7ecd-e111-bb59-0022644237b5