emulators/mupen64plus-video-z64: preserve a bit longer
Most of the mupen64plus ports support stage out of the box, just
remove "NO_STAGE=yes".  However, this port is broken for another
reason.  To avoid it being pruned for being unstaged, remove NO_STAGE
but leave it broken to give somebody a little more time to fix the
glew dependency issue.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@367594 35697150-7ecd-e111-bb59-0022644237b5