science: Fix mpqc, libint and unbreak ghemical, and libghemical atomically
science/ghemical would not link because libghemical.so (from science/
libghemical port) had never been properly linked.  Links to all mpqc "SC"
libraries were added to LDFLAGS to rectify this.  The configure breakage
and solution is described below.

After the version of lang/gcc was bumped from 4.7 to 4.8,
science/libghemical ceased to configure and it was marked broken.  After
recreating the conftest, it was discovered that two versions incompatible
versions of libgcc_s.so were getting pulled in by the realtime linker:
the base version and the gcc48 ports version.

The base version was getting pulled in by science/libint.  To unbreak
libghemical, libint is now built with lang/gcc.  It was necessary to
force libtool to link with LDFLAGS that Mk/bsd.gcc.mk sets so that
the runpaths match across libraries used by libghemical.

When science/mpqc was staged, it utilized libtool which renumbered all
the library versions from 8.0.1 to 7.1.0.  This was caused by the age
component being greater than 0.  By patching configure.in with a new
version, we can generate major SHLIB of 8 again.  While here, fix the
bin/sc-config tool to remove a bad include cflag.

With this fixes, science/ghemical builds successfully.  Bump all 4 of
these ports, remove any BROKEN designation and remove redundant
@dirrm in pkg-plist


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@369236 35697150-7ecd-e111-bb59-0022644237b5