java/java3d: Set to dual GPLv2 and MIT licensing
PR:		193186
Submitted by:	pfg@

After examining the LICENSE files in the existing tarball, Pedro came to
the conclusion that the license is a combination between GPLv2 and some
MIT-like licenses.  The package is redistributable; the outdated
restriction is probably left over from a previous version, so it has
been removed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@369295 35697150-7ecd-e111-bb59-0022644237b5