emulators/mame: clean up, remove zip dependency, support DragonFly
This port has a post-extract target that wants to use native unzip
on an extracted file.  This is an issue for DragonFly, but tar can
handle the zip file just file on both platforms, so switch to that.

DragonFly still has the incomplete FILE that it inherited from FreeBSD,
which is really a pain when FILE is a parameter in c++ routines.  This
adds an extra patch for DragonFly to use an alternative default logging
callbacks to all mame to finally build on this platform.

While here:
  - remove use of ${PORTNAME} where "mame" is the only correct value
  - use ".zip" instead of extract variable
  - Unmask installation commands
  - Group mkdir commands into one
  - Use NOT_FOR_ARCHS instead of conditional BROKEN
  - Convert to USES=python
  - Remove unnecessary conversion of -lpthread => -pthread


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370047 35697150-7ecd-e111-bb59-0022644237b5