net/hostapd: Update WWW + MASTER_SITES and support DragonFly
The domain for hostapd has changed from hostap.epitest.fi to w1.fi
although the former still redirects.  Update WWW and MASTER_SITES to
reflect the new name.

Regenerate the l2 packet patch so that hostapd also builds on DragonFly
(no-op for FreeBSD).

While here, rearrange makefile to remove need for <pre> and <post> and
use of $PORTNAME in $WRKSRC which would break if PORTNAME changes.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370693 35697150-7ecd-e111-bb59-0022644237b5