devel/matreshka: Fix build on i386 platforms
The author of matreshka informed me that GPRBuild 2014 has a bug in
it that suppresses flags like -msse2 and -march.  It's been fixed in the
Pro version, but not the GPL version.  He worked around the bug instead,
and this commit brings in the essential part of that workaround.  The
GPRBuild bug prevented Matreshka from building on i386 on all releases
and all platforms.

While here, remove @dirrm from pkg-plist, strip the installed libraries,
convert to USES=pgsql and tighten the makefile up a bit.

MFH:		2014Q4


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370966 35697150-7ecd-e111-bb59-0022644237b5