lang/gnat_util: Move PORTREVISION to central location
Many of the ports based on gcc-aux sources have a central location
defining their PORTREVISION to make it easy to reset all them all when
the base version is updated.  lang/gnat_util should have had, but did not
and as a result the PORTREVISION didn't get reset at all.  Centralize
it for the future.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371323 35697150-7ecd-e111-bb59-0022644237b5