New port: sysutils/backuppc-devel
BackupPC is a high-performance, enterprise-grade system for backing up
Linux, WinXX and MacOSX PCs and laptops to a server's disk.
BackupPC is highly configurable and easy to install and maintain.

WWW: http://backuppc.sourceforge.net

This is Version 4.0.0alpha3, please use with caution. On the other
hand, it is 10 month old, and the submitter is also the maintainer of
sysutils/backuppc.

PR:		183241
Submitted by:	Alexander Moisseev <moiseev@mezonplus.ru>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371400 35697150-7ecd-e111-bb59-0022644237b5