- Release ports maintained by games@
games@ team is quite small and inactive, so ports currently assigned
to it cannot be assumed as properly maintained. To not cause confusion
by making it look otherwise, and also to allow and promote adoption
by individual maintainers, release these ports back to the pool. Note
that this does not change the fact that games@ team still takes care
of these ports to excent of its capabilities.

Suggested by:	marino


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371872 35697150-7ecd-e111-bb59-0022644237b5