Add new port net-mgmt/seafile-gui
PR:		193312
Submitted by:	Jingfeng Yan

Seafile is a next-generation open source cloud storage system with
advanced support for file syncing, privacy protection and teamwork.
Collections of files are called libraries, and each library can be synced
separately. A library can be encrypted with a user chosen password. This
password is not stored on the server, so even the server admin cannot
view a file's contents.

Seafile allows users to create groups with file syncing, wiki, and
discussion to enable easy collaboration around documents within a team.
This is the desktop client for Seafile.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372230 35697150-7ecd-e111-bb59-0022644237b5