Add new port sysutils/rcm
PR:		194248
Submitted by:	Leonardo Brondani Schenkel

The rcm suite of tools is for managing dotfiles directories. This is a
directory containing all the .*rc files in your home directory (.zshrc,
.vimrc, and so on). These files have gone by many names in history, such as
"rc files" because they typically end in rc or "dotfiles" because they begin
with a period.

This suite is useful for committing your rc files to a central repository
to share, but it also scales to a more complex situation such as multiple
source directories shared between computers with some host-specific or
task-specific files.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372360 35697150-7ecd-e111-bb59-0022644237b5