www/p5-CGI-Application-Plugin-TT: make p5-CGI-Application-Plugin-DevHelp optional
Currently, p5-CGI-Application-Plugin-DevHelp is a hard dependency in
p5-CGI-Application-Plugin-TT. However, C::A::P::TT operates just fine
without it. The attached patch turns the dependency into an OPTION
that defaults to on.

Just for the sake of having a reference,
http://cpansearch.perl.org/src/CEESHEK/CGI-Application-Plugin-TT-1.05/
META.yml lists it as a "recommends" but not a "requires."

PR:		196472
Differential Revision:	https://reviews.freebsd.org/D1451
Submitted by:	adamw
Approved by:	marino(mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@376496 35697150-7ecd-e111-bb59-0022644237b5