lang/fpc: Prepare for 3.0 and the use of older bootstrap
The Free Pascal Compiler port currently mandates that the bootstrap
compiler be the same version of the compiler being built.  This means that
every update requires a new bootstrap compiler to be generated before the
update.  This goes against the philosophy of a bootstrap compiler, which
aims to allow older compilers to build self-hosting modern compilers.  FPC
is capable of this, but the feature wasn't being leveraged.

There are a lot of changes in this commit, but I believe the resulting
compiler is the same as before the commit, thus PORTREVISION has not been
bumped.  Changes include:

  * Add DragonFly bootstrap compiler 2.6.4 to list of distfiles
  * use OPSYS to make builds generic between FreeBSD and DragonFly
  * Add FREEBSD PLIST variable as FreeBSD builds an extra unit
  * Create the framework to have a bootstrap compiler of an older version
    than the current version (BOOTVER variable)
  * split PP variable out of MAKE_ARGS and use NEWPPC or BOOTPPC to
    specify PP uniquely (required when bootstrap version is different)
  * wrap compound command with parenthesis
  * Use options lists on TAR man page (xfz without hyphen is legacy)
  * unmask installation commands
  * respect 80-column limits, wrapping as necessary
  * bring in some dports-specific changes that have no effect on port
    to eliminate dports patch files

future work:

  When version 3.0 is released, the new fpc make program will be built
  which will have some moderate impacts to the build process and it
  will use a "bootstrap" target.

  The intention is to continue using the 2.6.4 bootstrap compilers to
  build the upcoming 3.0 compiler when it is released.  New bootstrap
  compilers will not be generated until the existing bootstrap compilers
  are unable to bootstrap the upcoming release.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@376655 35697150-7ecd-e111-bb59-0022644237b5