lang/fpc: Remove unnecessarily variable substition and EXTRACT_SUFX
There were several substitions like "${XXX:S/$/.something/} which is the
equivalent of appending, e.g. "${XXX}.something".

So for the sake of comprehension, change these to appends, and also just
hardcode ".tar.gz" instead of using ${EXTRACT_SUFX} for the same reason.
Since the distfiles are being manually defined, there's no real purpose
to have a variable in the name, and if the EXTRACT_SUFX changes for any
reason (say a new distfile with a different suffix requiring USES+=tar),
then the code actually breaks.  These changes make the code easier to
read, comprehend, and arguably makes the port more robust.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@376710 35697150-7ecd-e111-bb59-0022644237b5