Fix circular dep issues so this script can DTRT at boot.
This script needs to run after ldconfig, however it's deps cause:

a) A circular dependancy in rcorder
b) It starts before ldconfig

This causes the VBoxManage command to bail out when it's run by
this script because it can't find it's required libs.  The consequence of
this is host only networks aren't set up properly at boot.

By removing the explicit dependancy on ldconfig we get rid of the circular
dependancy issues.  However since the VBoxManage command does actually need
libraries in /usr/local/lib we set that explicitly with LD_LIBRARY_PATH.

There's not really a good solution to this issue.  The real "problem" is that /usr
could be an NFS mount and you have to run ldconfig after /usr is mounted. While
LD_LIBRARY_PATH is kind of ugly, and removing an explicit dependancy is ugly,
the other alternatives are uglier. (Break NFS mounted /usr for example)

I'd also argue that having this just not work (which is the state of it before
this commit) is pretty ugly too.

Also, I apologize in advance for not discussing this on emulation@ and just doing
a drive by commit, however hopefully the fact that "This fixes obvious breakage"
makes up for the fact that I skipped peer review, proper approval, and most importantly
giving the community to bikeshed this ad naseum.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@378074 35697150-7ecd-e111-bb59-0022644237b5