[NEW] devel/py-pytrie: Pure Python implementation of the trie data structure
A trie is an ordered tree data structure that is used to store a mapping
where the keys are sequences, usually strings over an alphabet. In
addition to implementing the mapping interface, tries allow finding the
items for a given prefix, and vice versa, finding the items whose keys
are prefixes of a given key.

WWW: https://bitbucket.org/gsakkis/pytrie


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@378925 35697150-7ecd-e111-bb59-0022644237b5